Develop

+++ Engine/Core/Components/Game/OwnerIdComponent.cs
@@ -1,10 +1,10 @@
-﻿using Entitas;
+﻿using Entitas;             
 
 namespace Lockstep.Core.Components.Game
 {
     [Game]
     public class OwnerIdComponent : IComponent
     {                        
-        public int value;
+        public byte value;
     }
 }
