+++ Engine/Core/Components/Game/TeamComponent.cs
@@ -1,10 +1,10 @@
-﻿using Entitas;
+﻿using Entitas;              
 
 namespace Lockstep.Core.Components.Game
 {
     [Game]
     public class TeamComponent : IComponent
     {
-        public int value;
+        public byte value;
     }
 }
